--- Smart-Contract-Verifier-Server/src/main/kotlin/com/savannasolutions/SmartContractVerifierServer/model/MessageStatus.kt
@@ -1,3 +1,0 @@
-package com.savannasolutions.SmartContractVerifierServer.model
-
-data class MessageStatus()
