+++ Smart-Contract-Verifier-Server/src/main/kotlin/com/savannasolutions/SmartContractVerifierServer/models/Agreements.kt
@@ -1,0 +1,3 @@
+package com.savannasolutions.SmartContractVerifierServer.models
+
+data class Agreements()
